requirements.txt
aiogram>=3.7.0
SQLAlchemy>=2.0.30
aiosqlite>=0.19.0
aiohttp>=3.9.5
APScheduler>=3.10.4
python-dotenv>=1.0.1
pydantic>=2.7,<3.0
loguru>=0.7.2
pytest>=8.2.0
pytest-asyncio>=0.23.7
